requirements.txt
requests
ollama
paho-mqtt
streamlit
pandas
python-dotenv
